Create BookHotel.php
This script processes the form submission for booking a hotel. It should include the Booking and BookingManager classes to create booking instances and manage them. This script is where you:

Collect input from the form.
Validate the input data.
Create a Booking object.
Add the Booking object to the collection managed by BookingManager.
Handle errors and provide feedback to the user.